split schema from main plugin script
