+++ README.md
@@ -1,11 +1,6 @@
-# Python Plugin Template Project
+# Python Plugin Template
 
-## Image Building
-
-You can change this plugin's image version tag in
-`.github/workflows/carpenter.yaml` by editing the
-`IMAGE_TAG` variable, and pushing that change to the
-branch designated in that workflow.
+This plugin serves as a template from which new plugins can be built.
 
 # Autogenerated Input/Output Documentation by Arcaflow-Docsgen Below
