+++ docker-compose.yaml
@@ -4,6 +4,6 @@ services:
     image: quay.io/arcalot/arcaflow-plugin-template-python
     build: .
     volumes:
-      - source: ./example.yaml
+      - source: ./configs/example.yaml
         target: /config/example.yaml
         type: bind
