+++ pyproject.toml
@@ -6,7 +6,7 @@ authors = ["Arcalot"]
 license = "Apache-2.0+GPL-2.0-only"
 
 packages = [
-   { include="arcaflow_plugin_template_python.py", from="./arcaflow_plugin_template_python"  },
+   { include="template_python_plugin.py", from="./arcaflow_plugin_template_python"  },
 ]
 
 [tool.poetry.dependencies]
